Update LogitWithR.r
The code for writing the coefficient dbf table was not doing what you thought it was doing.  Compare the results to the output from the command:

summary(fit)

This may break the corresponding part in the python file, but I am still new to python and lack the chops to correct it.