Create GLMwithR.r
This script is an adaptation of LogitWithR.r.  It runs a negative binomial regression, prints a series of diagnostic graphs into the working directory, and returns the regression results to ArcGIS as a shapefile and two .dbf tables showing coefficient data and information criteria for different combinations of independent variables in the model.